+++ inductiva/api/__init__.py
@@ -1,5 +1,5 @@
 """Functionality to interact with the Web API."""
-from .methods import invoke_api, invoke_api_from_fn_ptr, run_simulation, run_async_simulation
+from .methods import invoke_api, invoke_api_from_fn_ptr
 from .methods import download_output, invoke_async_api, get_task_status
 from .methods import validate_api_key
 from .methods import get_client
