+++ inductiva/tasks/__init__.py
@@ -1,3 +1,4 @@
 #pylint: disable=missing-module-docstring
-from .methods import fetch_task_output, get_task_status, get_task_info, get_tasks_info
 from .task import Task
+from .methods import fetch_task_output, get_task_status, get_task_info, get_tasks_info
+from .run_simulation import run_simulation, run_async_simulation
